+++ src/main/java/org/example/dodalock/dodalock/DodaLockMain.java
@@ -2,12 +2,10 @@ package org.example.dodalock.dodalock;
 
 import org.bukkit.Location;
 import org.bukkit.plugin.java.JavaPlugin;
+import org.example.dodalock.dodalock.commands.GlobalCommands;
 import org.example.dodalock.dodalock.gui.GuiItemsManager;
 import org.example.dodalock.dodalock.items.ItemsManager;
-import org.example.dodalock.dodalock.listeneres.BunchKeysListener;
-import org.example.dodalock.dodalock.listeneres.LocksListener;
-import org.example.dodalock.dodalock.listeneres.MenuListener;
-import org.example.dodalock.dodalock.listeneres.PlayerJoinListener;
+import org.example.dodalock.dodalock.listeneres.*;
 import org.example.dodalock.dodalock.tasks.ClearBunchKeysInventory;
 import org.example.dodalock.dodalock.utils.FormattableUtils;
 import org.example.dodalock.dodalock.utils.WorldUtils;
@@ -45,10 +43,11 @@ public final class DodaLockMain extends JavaPlugin {
         this.getServer().getPluginManager().registerEvents(new LocksListener(), this);
         this.getServer().getPluginManager().registerEvents(new BunchKeysListener(), this);
         this.getServer().getPluginManager().registerEvents(new PlayerJoinListener(), this);
+        //this.getServer().getPluginManager().registerEvents(new KeyCloneCraftListener(), this);
     }
 
     private void registerCommands() {
-
+        new GlobalCommands(this);
     }
 
     private void checkLocks() {