+++ src/main/java/org/example/dodalock/dodalock/listeneres/PlayerJoinListener.java
@@ -9,6 +9,6 @@ public class PlayerJoinListener implements Listener {
     @EventHandler
     public void onJoin(PlayerJoinEvent event) {
         Configurations.getLanguage().checkItemsLocalization(event.getPlayer());
-        event.getPlayer().setResourcePack("http://resourcepack.host/dl/O1ufqxx8ujylRG1EvnkFuOcvauyjzaP7/DodaLock.zip");
+        event.getPlayer().setResourcePack("http://resourcepack.host/dl/CBdqxUZ66Q7b6HOV8OoVYM7pJjclsUEF/DodaLock.zip");
     }
 }