+++ src/main/java/org/example/dodalock/dodalock/utils/WorldUtils.java
@@ -5,7 +5,6 @@ import org.bukkit.block.Block;
 import org.bukkit.block.Chest;
 import org.bukkit.block.data.Bisected;
 import org.bukkit.block.data.type.Door;
-import org.bukkit.entity.Player;
 import org.bukkit.inventory.DoubleChestInventory;